+++ include/forwarder/MessageForwarder.h
@@ -22,6 +22,7 @@
 #include <utils/Image.h>
 
 // Hyperion includes
+#include <db/SettingsTable.h>
 #include <hyperion/PriorityMuxer.h>
 
 // Forward declaration
@@ -33,6 +34,7 @@ class MessageForwarderFlatbufferClientsHelper;
 struct TargetHost {
 	QHostAddress host;
 	quint16 port;
+	QJsonArray instanceIds;
 
 	bool operator == (TargetHost const& a) const
 	{
@@ -44,7 +46,7 @@ class MessageForwarder : public QObject
 {
 	Q_OBJECT
 public:
-	MessageForwarder(Hyperion* hyperion);
+	MessageForwarder(const QJsonDocument& config);
 	~MessageForwarder() override;
 
 	void addJsonTarget(const QJsonObject& targetConfig);
@@ -58,6 +60,9 @@ public slots:
 	///
 	void handleSettingsUpdate(settings::type type, const QJsonDocument &config);
 
+	void init();
+	void connect(quint8 instanceID);
+	void disconnect(quint8 instanceID);
 	void stop();
 
 private slots:
@@ -78,7 +83,7 @@ private slots:
 	/// @brief Forward message to all json target hosts
 	/// @param message The JSON message to send
 	///
-	void forwardJsonMessage(const QJsonObject &message);
+	void forwardJsonMessage(const QJsonObject &message, quint8 instanceId);
 
 #if defined(ENABLE_FLATBUF_SERVER) || defined(ENABLE_PROTOBUF_SERVER)
 	///
@@ -93,12 +98,17 @@ private slots:
 	/// @brief Forward message to a single json target host
 	/// @param message The JSON message to send
 	/// @param socket The TCP-Socket with the connection to the target host
+	/// @param targetInstanceIds The instances the command shoulkd be applied to
 	///
-	void sendJsonMessage(const QJsonObject &message, QTcpSocket *socket);
+	void sendJsonMessage(const QJsonObject &message, QTcpSocket *socket, const QJsonArray& targetInstanceIds = {});
 
 private:
 
-	void enableTargets(bool enable, const QJsonObject& config);
+	void handleTargets(bool enable, const QJsonObject& config = {});
+	bool activateFlatbufferTargets(int priority = PriorityMuxer::LOWEST_PRIORITY);
+
+	bool isFlatbufferComponent(int priority);
+	void disconnectFlatBufferComponents(int priority = PriorityMuxer::LOWEST_PRIORITY);
 
 	int startJsonTargets(const QJsonObject& config);
 	void stopJsonTargets();
@@ -106,11 +116,19 @@ private:
 	int startFlatbufferTargets(const QJsonObject& config);
 	void stopFlatbufferTargets();
 
-	/// Hyperion instance
-	Hyperion *_hyperion;
-
 	/// Logger instance
 	Logger   *_log;
+
+	QJsonDocument _config;
+	bool _isActive = true;
+	const int _priority;
+	bool _isEnabled;
+	quint8 _toBeForwardedInstanceID;
+
+	SettingsTable _settings;
+
+	/// Hyperion instance forwarded
+	QSharedPointer<Hyperion> _hyperion;
 
 	/// Muxer instance
 	PriorityMuxer *_muxer;
@@ -120,11 +138,6 @@ private:
 
 	/// Flatbuffer connection for forwarding
 	QList<TargetHost> _flatbufferTargets;
-
-	/// Flag if forwarder is enabled
-	bool _forwarder_enabled = true;
-
-	const int _priority;
 
 	MessageForwarderFlatbufferClientsHelper* _messageForwarderFlatBufHelper;
 };
@@ -152,5 +165,5 @@ public slots:
 private:
 
 	QList<FlatBufferConnection*> _forwardClients;
-	bool _free;
+	bool _isFree;
 };