+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for Skydimo devices
 - Support gaps on Matrix Layout (#1696)
 - Windows: Added a new grabber that uses the DXGI DDA (Desktop Duplication API). This has much better performance than the DX grabber as it does more of its work on the GPU.
+- Support to freely select source and target instances to be used by forwarder
 
 - Support to import, export and backup Hyperion's full configuration via the UI, JSON-API and commandline (`--importConfig, --exportConfig`) (#804)
 - Allow to force starting Hyperion in read-only mode (`--readonlyMode`)
@@ -46,6 +47,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactored: Database access layer
 - Refactored: Hyperion's configuration database is validated before start-up (and migrated, if required)
 - Refactored: Python to enable parallel effect processing under Python 3.12
+- Refactored: Forwarder
+- Refactored: Flatbuffer connection- and client handling
+- Standalone grabber do not capture screens when no connection to remote host
 - Fixed: Python 3.12 crashes (#1747)
 - osX Grabber: Use ScreenCaptureKit under macOS 15 and above
 - Removed maximum LED number constraint from Matrix layout schema which was not synced with the UI behaviour (#1804)